Split tests from Acceleration_test into dedicated files
